requirements.txt
Pillow==8.1.2
PyQt5==5.15.4
